Moved logging to background script, created new file usb-devices-logging
